perf(ci): move the scaffolder template matrix to the release boundary
Frequency policy for the packed pool: per-PR release gates keep the
single-cycle consumer proofs (pack once, install once per contract) plus one
full scaffolder journey — the minimal template, which covers the installed
scaffolder bin, template scaffold, scaffolder-driven npm install, project
check, and clean validate. The mcp-server and cli-tool template runs move to
scaffold-packed-matrix.e2e.test.ts, which runs in `test:packed:release`
(pre-publish check:release and a new nightly CI schedule), not per PR. The
shared scaffold fixture moves to tests/support/scaffold-fixture.ts; test
names are unchanged, so the release-boundary pool census is identical to the
old per-PR pool census, and the per-PR pool drops exactly the two matrix
templates. CI's release-gates job now runs check:release:ci (per-PR pool);
the nightly packed-matrix job runs the full check:release.
